ugs-core/src/com/willwinder/universalgcodesender/model/Direction.java
/*
    Copyright 2020 Will Winder

    This file is part of Universal Gcode Sender (UGS).

    UGS is free software: you can redistribute it and/or modify
    it under the terms of the GNU General Public License as published by
    the Free Software Foundation, either version 3 of the License, or
    (at your option) any later version.

    UGS is distributed in the hope that it will be useful,
    but WITHOUT ANY WARRANTY; without even the implied warranty of
    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
    GNU General Public License for more details.

    You should have received a copy of the GNU General Public License
    along with UGS.  If not, see <http://www.gnu.org/licenses/>.
 */
package com.willwinder.universalgcodesender.model;

import org.apache.commons.lang3.builder.EqualsBuilder;
import org.apache.commons.lang3.builder.HashCodeBuilder;
import org.apache.commons.lang3.builder.ToStringBuilder;

/**
 * Defines a direction using a double value between -1.0 to 1.0.
 * Zero means that it should not move that axis.
 *
 * @author Joacim Breiler
 */
public class Direction {
    private final Double x;
    private final Double y;
    private final Double z;

    public Direction(Double x, Double y, Double z) {
        this.x = parseDirectionValue(x);
        this.y = parseDirectionValue(y);
        this.z = parseDirectionValue(z);
    }

    private double parseDirectionValue(Double value) {
        if (value == null) {
            return 0;
        }

        return Math.max(Math.min(value, 1.0d), -1.0d);
    }

    public Double getX() {
        return x;
    }

    public Double getY() {
        return y;
    }

    public Double getZ() {
        return z;
    }

    @Override
    public boolean equals(Object o) {
        return EqualsBuilder.reflectionEquals(this, o);
    }

    @Override
    public int hashCode() {
        return HashCodeBuilder.reflectionHashCode(this);
    }

    @Override
    public String toString() {
        return ToStringBuilder.reflectionToString(this);
    }
}
